+++ HISTORY.md
@@ -1,6 +1,7 @@
 # History
 
 # 0.7.0 (dev)
+- Add counts function to show the number of total and unique values in a column
 
 # 0.6.0 (2020-7-4)
 - Frequency results were not proper percentages. Updated so percentages used correctly.